allow custom roles with client credentials authentication flow
Drop the requirement to have custom roles as role claim in the jwt and the header. This allows to authenticate with EntraID tokens acquired via client credentials flow. EntraID does not allow to set custom claims to be set on request. For dab to switch into a custom role the header with the custom role and a valid jwt token must be present.